Fix Makefile
I am unsure as to why, but the dependency expressed as "mcrouter: deps"
doesn't seem to work correctly.

When invoking the equivalent of "make mcrouter", I expect it to build
all the transitive dependencies of "deps", but it only builds the first
dependency that it determines it needs to build, and then exits with
a return value of "2".

I'm just reverting this to the previous style that is redundant but
works correctly.