"git clean" during checkout of dependency
This repo defines "recipes" for dependencies which are essentially build
scripts.  Some of the build systems used (or at least the way they are
used) will cache some state between invocations, which is confusing when
testing different values for things like "LDFLAGS" and
"LD_LIBRARY_PATH".  Specifically, the cmake based build scripts.

This change ensures that when force re-running a given recipe, we do a
full clean of the repo.

n.b.: we are *not* cleaning the target install dir.