Build jemalloc from source.
This introduces "jemalloc" as a library dependency that we build from
source.

While this library is available from our runtime system's package
respository, mcrouter would log an error at startup implying that there
is a page size mis-match between the current kernel, and the system on
which jemalloc was compiled.

As this package available via the EPEL[1] and not Amazon Linux directly,
this isn't entirely surprising.

Compiling "jemalloc" is very straightforward, but getting all the
dependencies to correctly link against it during their builds was a bit
challenging.  Ultimately I made the following changes:

1. The default value we set for *all* recipes is:

        LDFLAGS="-L$INSTALL_DIR/lib -ljemalloc

- "-ljemalloc" means "link against the 'jemalloc'" library.
- "-L$INSTALL_DIR/lib" search "$INSTALL_DIR/lib" when looking for
libraries.

2. I removed redundant LDFLAGS from each recipe following the change to
the default.

3. Some builds, notable those that produce binary executables (not just
shared libraries), also need to include "-Wl,-rpath=$INSTALL_DIR/lib",
which embeds metadata about the dynamic libraries' location into the
executable.  Conceptually, this doesn't make sense to me, as we are
expecting users to manage their own library path, but not setting this
causes the builds to fail, and I have a limited appetite for debugging
and understanding these build tools.

[1]: https://fedoraproject.org/wiki/EPEL