Refine binary archive.
This streamlines the resulting archive in three ways:

1. It only includes mcrouter and mcpiper from /bin (no thrift, zfmt,
etc)
2. It only includes the libraries that mcrouter and mcpiper actually
link against.  This removes a bunch of boost libraries from the archive.
3. It includes the dynamic libraries not only from INSTALL_DIR, but
also from the host system (e.g. libssl and libz).

This archive is *NOT PORTABLE* in any meaningful way, it is only
meant to run on essentially the same target system, amazon linux 2.